requirements.txt
# Core dependencies
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0

# Testing dependencies
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.11.1
factory-boy==3.3.0

# Development dependencies
black==23.7.0
flake8==6.0.0
mypy==1.5.1
pre-commit==3.4.0

# Logging
structlog==23.1.0

# Validation
pydantic==2.3.0
marshmallow==3.20.1
